+++ NEWS.rst
@@ -4,6 +4,10 @@ NEWS for system-image updater
 
 0.9 (2013-XX-XX)
 ================
+ * Rename DBus method IsUpdateAvailable() to CheckForUpdate() and make it
+   asynchronous.  Rename the UpdatePending() signal to UpdateAvailableStatus()
+   and have it contain a boolean flag which indicates whether an update is
+   available or not.  Make GetUpdate() actually asynchronous.  (LP: #1204976)
 
 0.8 (2013-07-24)
 ================